// src/enum/order.enum.ts
export enum OrderStatus {
  IN_CART = 'IN_CART',
  ORDER_PLACED = 'ORDER_PLACED',
  PACKING = 'PACKING',
  PACKED = 'PACKED',
  DELIVERING = 'DELIVERING',
  DELIVERED = 'DELIVERED',
  PRODUCT_PROBLEMS = 'PRODUCT_PROBLEMS',
  WAITING_FOR_PAYMENT = 'WAITING_FOR_PAYMENT',
  WAITING_FOR_APPROVED = 'WAITING_FOR_APPROVED',
  CLAIM = 'CLAIM',
  SUCCESS = 'SUCCESS',
  CANCELED = 'CANCELED',
}